# cti-hub/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GRAMMAR Orchestrator API",
    "description": "API for routing natural language prompts to GRAMMAR's Technical Language Index, First-Draft Harness, and Code Execution engine.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://76.13.96.107:3001"
    }
  ],
  "paths": {
    "/api/grammar/execute": {
      "post": {
        "summary": "Execute grammatical logic or app creation",
        "description": "Routes user prompt to GRAMMAR's orchestrator, invoking Technical Language Index (TLI), Look/Listen/Learn (LLL), and First-Draft Harness (FDH).",
        "operationId": "executeGrammarTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string",
                    "description": "The natural language instruction from the user."
                  },
                  "sessionId": {
                    "type": "string",
                    "description": "Optional session ID to maintain chat context."
                  }
                },
                "required": ["prompt"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful execution",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "plan": { 
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "result": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}